[feat] : #41 timer 모듈분리
