+++ feature/clip/src/main/res/layout/item_clip_clip.xml
