+++ feature/clip/src/main/res/layout/item_clip_detail_link.xml
