+++ feature/clip/src/main/res/layout/item_clip_edit_clip.xml
