+++ feature/timer/src/main/java/org/sopt/timer/CompleteTimerViewHolder.kt
@@ -1,8 +1,8 @@
-package org.sopt.mainfeature.timer
+package org.sopt.timer
 
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.mainfeature.databinding.ItemTimerCompleteBinding
-import org.sopt.mainfeature.timer.dummymodel.Timer
+import org.sopt.timer.databinding.ItemTimerCompleteBinding
+import org.sopt.timer.dummymodel.Timer
 
 class CompleteTimerViewHolder(
   private val binding: ItemTimerCompleteBinding,
