+++ feature/timer/src/main/java/org/sopt/timer/dummymodel/Timer.kt
@@ -1,4 +1,4 @@
-package org.sopt.mainfeature.timer.dummymodel
+package org.sopt.timer.dummymodel
 
 data class Timer(
   val id: Int,
